Random Customers
Start of  the structure that will create a list of customers the player might encounter for an order. They have a set spawn location for now, and it spawns in 1 of 3 testing sprites (not final) at the start of the game. To build on this, they will have randomized orders, with some especially unique characters.