Added sound, fixed cauldron color
The daytime now has ambient noise. In addition to that, the cauldron bubbles ambiently, gets louder with each ingredient (to a point), and now makes a sound when potions are completed. Lastly, fixed an issue where the cauldron color would not reset to the default blue upon making a potion.