Fixed clock, Added sprite addon, Book UI (#116)
* Fixed the clock and added Scene Manager

Fixed the clock so it properly shows how long a day lasts and added a rough version of the scene manager that still needs to be refactored to fit our project

* Shader attempts, Book UI

Tried to implement shader that would allow for dynamic liquid within the potion bottles. Didn't end up working, but will continue on that in the future. Trying to also get the ingredient book ui to also work

* Added a sprite addon and implemented animation player