Fixed potion pedestal tweening bug
Previously, the potion pedestal would not tween a potion to the center seemingly at random. I found this to be caused by taking the potion off of the pedestal and putting it back on before letting go of it again. To fix this, the body of the potion is removed from the heldNodes list, meaning it will be subjected to the same code as a potion entering for the first time.