Changed Ingredient Book
The Ingredient book now loads all the ingredients from areas you have visited instead of just having them hard coded in. The book on the shelf in main now also disappears when clicked on and I added an X to the ingredient that the player has to click on to close it. (There is a small dead zone on the X button and I don't know what is causing it)